implementing a basic image cropper

+++ frontend/components/post/Bookmark.vue
@@ -1,13 +1,11 @@
 <template>
   <div>
-    <UButton
-      :icon="Icon"
+    <button
       @click="handleBookmark"
-      size="sm"
-      color="primary"
-      square
-      variant="none"
-    />
+      class="h-7 w-7 flex items-center justify-center text-gray-700 rounded-full hover:bg-gray-100"
+    >
+      <Icon :name="saveIcon" size="24" />
+    </button>
   </div>
 </template>
 
@@ -29,48 +27,52 @@ const props = defineProps({
 
 const saved = ref(props.saved);
 
-const Icon = computed(() => {
+const saveIcon = computed(() => {
   if (saved.value) return "i-material-symbols-bookmark-added-rounded";
   else return "i-material-symbols-bookmark-add-outline-rounded";
 });
 
 const handleBookmark = async () => {
-  if (!saved.value) {
-    try {
-      const res = await $fetch(
-        "http://127.0.0.1:8000/api/users/bookmarks/add/",
-        {
-          method: "post",
-          headers: {
-            Authorization: `Bearer ${auth.accessToken}`,
+  if (auth.user) {
+    if (!saved.value) {
+      try {
+        const res = await $fetch(
+          "http://127.0.0.1:8000/api/users/bookmarks/add/",
+          {
+            method: "post",
+            headers: {
+              Authorization: `Bearer ${auth.accessToken}`,
+            },
+            body: {
+              post_id: props.postId,
+            },
           },
-          body: {
-            post_id: props.postId,
+        );
+        saved.value = true;
+      } catch (err) {
+        throw new Error(err);
+      }
+    } else {
+      try {
+        const res = await $fetch(
+          "http://127.0.0.1:8000/api/users/bookmarks/delete/",
+          {
+            method: "delete",
+            headers: {
+              Authorization: `Bearer ${auth.accessToken}`,
+            },
+            body: {
+              post_id: props.postId,
+            },
           },
-        },
-      );
-      saved.value = true;
-    } catch (err) {
-      throw new Error(err);
+        );
+        saved.value = false;
+      } catch (err) {
+        throw new Error(err);
+      }
     }
   } else {
-    try {
-      const res = await $fetch(
-        "http://127.0.0.1:8000/api/users/bookmarks/delete/",
-        {
-          method: "delete",
-          headers: {
-            Authorization: `Bearer ${auth.accessToken}`,
-          },
-          body: {
-            post_id: props.postId,
-          },
-        },
-      );
-      saved.value = false;
-    } catch (err) {
-      throw new Error(err);
-    }
+    navigateTo("/auth/login");
   }
 };
 </script>
